Fix same_mask casts

+++ schemes/mceliece6960119f/vec/pk_gen.c
@@ -123,7 +123,7 @@ static int mov_columns(uint64_t mat[][ (SYS_N + 63) / 64 ], uint32_t *perm) {
     for (j = 0;   j < 32; j++)
         for (k = j + 1; k < 64; k++) {
             d = perm[ row + j ] ^ perm[ row + k ];
-            d &= same_mask(k, ctz_list[j]);
+            d &= same_mask((uint16_t)k, (uint16_t)ctz_list[j]);
             perm[ row + j ] ^= d;
             perm[ row + k ] ^= d;
         }
@@ -141,7 +141,7 @@ static int mov_columns(uint64_t mat[][ (SYS_N + 63) / 64 ], uint32_t *perm) {
         for (j = 0; j < 32; j++)
             for (k = j + 1; k < 64; k++) {
                 d = buf[ j ] ^ buf[ k ];
-                d &= same_mask(k, ctz_list[j]);
+                d &= same_mask((uint16_t)k, (uint16_t)ctz_list[j]);
                 buf[ j ] ^= d;
                 buf[ k ] ^= d;
             }
